fix(types): relational WhereInput types causes mypy SIGKILL
this was "fixed" by broadening the relational filter sub-type to
Dict[str, Any] when not targeting fully recursive types.

Sacrificing the type safety of this is is something that I am okay with
as full type safety can still be achieved by using pyright instead and
the mypy limitation has been documented.

this commit also does the following:

- splits up the typesafety tests to different environments
for different type checkers as it was frustrating how long the mypy
typesafety tests would take

- upgrades the mypy version we type check against to 0.910