refactor(http): use httpx by default and remove other libraries
The main reason for this change is to improve user experience and remove
any ambiguity with config options.

Previously we would require that users install the client with an extra,
for example:

pip install prisma-client[aiohttp]

This is not a great experience, users installing the client for the
first time should not have to decide what http library will be used
internally. Switching to just httpx means that the client can now be
installed without any extras.

pip install prisma-client

There was also ambiguity with the http config option, it was not clear
that this option controls whether or not the generated client is
asynchronous or not.