fix: use arm64 prisma binaries on arm64 machines
Fetch arm64 binaries when running on an arm64 machine.

This allows prisma-client-py to work on Linux and macOS machines with
arm64 CPUs (it actually already worked on macOS because of Rosetta but
this lets it use native binaries.)

Throw a RuntimeError if the machine is not arm64 or x86_64 to avoid
downloading binaries with the wrong architecture.