add TYPE_CHECKING check back for StrEnum definition
As per the discussion in the related pull request, I have added back the TYPE_CHECKING check to prevent issues with linters for the try-except import. I have left the try-except in as-per the discussion, but I am unsure of the benefit to having both the check and the try-except.